// src/app/Interfaces/product_category_interface.ts
export interface IProductCategory {
  id: string;

  PCName: string;

  PCMetaDataDescription: string;

  PCParentCategoryId: string;

  PCDescription: string;

  PCNewURL: string;

  OrganiztionId: string;

  PCMetaDataTitle: string;

  WebSiteName: string;

  PCMetaDataKeywords: string;

  PCOldURL: string;

  access_token: string;
  user_id: string;
}